user experience survey

+++ src/BlendedConcept/Survey/Application/Repositories/Eloquent/SurveyRepository.php
@@ -19,14 +19,16 @@ class SurveyRepository implements SurveyRepositoryInterface
 {
     public function getUserExperienceSurveyList($filters = [])
     {
-        $surveys = SurveyResource::collection(SurveyEloquentModel::filter($filters)->where('type', 'USERREXP')->orderBy('id', 'desc')->paginate($filters['perPage'] ?? 10));
+        $surveys = SurveyResource::collection(SurveyEloquentModel::filter($filters)->where('type', 'USEREXP')->orderBy('id', 'desc')->paginate($filters['perPage'] ?? 10));
 
         return $surveys;
     }
 
     public function showSurvey($id)
     {
-        $survey = new SurveyResource(SurveyEloquentModel::with('questions.options')->find($id));
+        $survey = new SurveyResource(SurveyEloquentModel::with(['questions' => function ($query) {
+            $query->orderBy('order', 'asc'); // Replace 'your_question_column' with the actual column name in the questions table
+        }, 'questions.options'])->find($id));
 
         return $survey;
     }
@@ -42,11 +44,12 @@ class SurveyRepository implements SurveyRepositoryInterface
 
             $questions = json_decode($request->questions, true);
 
-            foreach ($questions as $question) {
+            foreach ($questions as $key => $question) {
                 $questionArray = [
                     'survey_id' => $survey_id,
                     'question_type' => $question['question_type'],
                     'question' => $question['question'],
+                    'order' => $key
                 ];
                 $questionRequest = QuestionMapper::fromRequest($questionArray);
                 $questionEloquent = QuestionMapper::toEloquent($questionRequest);
@@ -54,16 +57,18 @@ class SurveyRepository implements SurveyRepositoryInterface
 
                 $question_id = $questionEloquent->id;
 
-                $options = $question['options'];
-                foreach ($options as $option) {
-                    if ($option !== '') {
-                        $optionArray = [
-                            'question_id' => $question_id,
-                            'content' => $option,
-                        ];
-                        $optionRequest = QuestionOptionMapper::fromRequest($optionArray);
-                        $questionOptionEloquent = QuestionOptionMapper::toEloquent($optionRequest);
-                        $questionOptionEloquent->save();
+                if($questionEloquent->question_type != 'SHORT_ANSWER'){
+                    $options = $question['options'];
+                    foreach ($options as $option) {
+                        if ($option !== '') {
+                            $optionArray = [
+                                'question_id' => $question_id,
+                                'content' => $option,
+                            ];
+                            $optionRequest = QuestionOptionMapper::fromRequest($optionArray);
+                            $questionOptionEloquent = QuestionOptionMapper::toEloquent($optionRequest);
+                            $questionOptionEloquent->save();
+                        }
                     }
                 }
             }
@@ -84,6 +89,13 @@ class SurveyRepository implements SurveyRepositoryInterface
             $surveyEloquent->fill($surveyArray);
             $surveyEloquent->save();
 
+            $questions = json_decode(request()->questions, true);
+            foreach($questions as $key => $question){
+                $questionEloquent = QuestionEloquentModel::find($question['id']);
+                $questionEloquent->order = $key;
+                $questionEloquent->update();
+            }
+
             DB::commit();
         } catch (\Exception $error) {
             DB::rollBack();
@@ -102,10 +114,13 @@ class SurveyRepository implements SurveyRepositoryInterface
         DB::beginTransaction();
 
         try {
+            $last_order = QuestionEloquentModel::where('survey_id', $request->survey_id)->orderBy('order', 'desc')->value('order');
+
             $questionRequest = QuestionMapper::fromRequest([
                 'survey_id' => $request->survey_id,
                 'question_type' => $request->question_type,
                 'question' => $request->question,
+                'order' => $last_order + 1
             ]);
 
             $questionEloquent = QuestionMapper::toEloquent($questionRequest);
@@ -113,17 +128,19 @@ class SurveyRepository implements SurveyRepositoryInterface
 
             $question_id = $questionEloquent->id;
 
-            $options = json_decode($request->options, true);
+            if($questionEloquent->question_type != 'SHORT_ANSWER'){
+                $options = json_decode($request->options, true);
 
-            foreach ($options as $option) {
-                if ($option !== '') {
-                    $optionArray = [
-                        'question_id' => $question_id,
-                        'content' => $option,
-                    ];
-                    $optionRequest = QuestionOptionMapper::fromRequest($optionArray);
-                    $questionOptionEloquent = QuestionOptionMapper::toEloquent($optionRequest);
-                    $questionOptionEloquent->save();
+                foreach ($options as $option) {
+                    if ($option !== '') {
+                        $optionArray = [
+                            'question_id' => $question_id,
+                            'content' => $option,
+                        ];
+                        $optionRequest = QuestionOptionMapper::fromRequest($optionArray);
+                        $questionOptionEloquent = QuestionOptionMapper::toEloquent($optionRequest);
+                        $questionOptionEloquent->save();
+                    }
                 }
             }
 
@@ -149,26 +166,29 @@ class SurveyRepository implements SurveyRepositoryInterface
             $questionEloquent->fill($questionArray);
             $questionEloquent->save();
 
-            $oldOptions = $questionEloquent->options;
+            if($questionEloquent->question_type != 'SHORT_ANSWER'){
+                $oldOptions = $questionEloquent->options;
 
-            foreach ($oldOptions as $oldOption) {
-                $oldOption->delete();
-            }
+                foreach ($oldOptions as $oldOption) {
+                    $oldOption->delete();
+                }
 
-            $question_id = $questionEloquent->id;
-            $newOptions = json_decode($questionArray['options'], true);
+                $question_id = $questionEloquent->id;
+                $newOptions = json_decode($questionArray['options'], true);
 
-            foreach ($newOptions as $option) {
-                if ($option !== '') {
-                    $optionArray = [
-                        'question_id' => $question_id,
-                        'content' => $option,
-                    ];
-                    $optionRequest = QuestionOptionMapper::fromRequest($optionArray);
-                    $questionOptionEloquent = QuestionOptionMapper::toEloquent($optionRequest);
-                    $questionOptionEloquent->save();
+                foreach ($newOptions as $option) {
+                    if ($option !== '') {
+                        $optionArray = [
+                            'question_id' => $question_id,
+                            'content' => $option,
+                        ];
+                        $optionRequest = QuestionOptionMapper::fromRequest($optionArray);
+                        $questionOptionEloquent = QuestionOptionMapper::toEloquent($optionRequest);
+                        $questionOptionEloquent->save();
+                    }
                 }
             }
+
             DB::commit();
         } catch (\Exception $error) {
             DB::rollBack();
